executor/requirements.txt
Flask
PyYAML
torch
sklearn
pillow
requests